requirements.txt
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
urllib3==2.3.0
